MVP: post form directly to Google Form
Had to change the submit button to an actual button in order for the
form validation to work.

Also, we can hook form.submit rather than button.click (so you can
submit using enter, for example)

Now we have a good fallback plan in case we have AJAX problems.

TODO:
- check the checkbox
- post with AJAX instead, redirect to thanks